requirements/base.txt
pytz==2019.2  # https://github.com/stub42/pytz
python-slugify==3.0.3  # https://github.com/un33k/python-slugify
Pillow==6.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.3.8  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
sentry-sdk==0.10.2 #https://github.com/getsentry/sentry-python
uuid==1.30


# Django
# ------------------------------------------------------------------------------
django==2.2.4  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-rest-swagger==2.2.0  # https://django-rest-swagger.readthedocs.io/en/latest/
djangorestframework-jwt==1.11.0  # http://jpadilla.github.io/django-rest-framework-jwt/
django-celery-beat==1.5.0 # https://github.com/celery
django-celery-results==1.1.2  # https://github.com/celery/django-celery-results
django-debug-toolbar==2.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.2.1  # https://github.com/django-extensions/django-extensions
prometheus-client==0.7.1  # https://github.com/korfuri/django-prometheus
django-prometheus==1.0.15  # https://github.com/korfuri/django-prometheus
django-cors-headers==3.1.0  # https://github.com/adamchainz/django-cors-headers
django-des==2.4.1  # https://github.com/jamiecounsell/django-des

# Django REST Framework
djangorestframework==3.10.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Para validação e formatação de CNPJ e CPF
brazilnum==0.8.8

# Para melhoria da aparência do Django Admin
# https://github.com/fabiocaccamo/django-admin-interface
django-admin-interface==0.11.1

# Pandas - Para importação de dados
# pandas==0.25.1

# Para gravação de log de alterações nos modelos
# https://django-auditlog.readthedocs.io/en/latest/index.html
django-auditlog
# Pegando última versão diretamente do repositório oficial.
#-e git://github.com/jjkester/django-auditlog.git##egg=django-auditlog

# Para filtros no DRF
# https://django-filter.readthedocs.io/en/master/#
django-filter==2.1.0

# Notification app for Django,
# https://github.com/django-notifications/django-notifications
django-notifications-hq==1.5.0

flower==0.9.3

# Testing
# ------------------------------------------------------------------------------
mypy==0.720  # https://github.com/python/mypy
pytest==5.1.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
pytest-watch==4.2.0
pytest-xdist==1.29.0
pytest-services==1.3.1
pytest-forked==1.0.2
pytest-django==3.5.1
pytest-coverage==0.0
pytest-cover==3.0.0 # https://github.com/pytest-dev/pytest-cov
model-mommy==1.6.0
freezegun==0.3.12 # https://github.com/spulec/freezegun

# Code quality
# ------------------------------------------------------------------------------
coverage==4.5.4  # https://github.com/nedbat/coveragepy
black==19.3b0  # https://github.com/ambv/black
pylint-django==2.0.11  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
isort==4.3.21

#Flake 8
# ------------------------------------------------------------------------------
flake8==3.7.8  # https://github.com/PyCQA/flake8
flake8-import-order==0.18.1 # https://github.com/PyCQA/flake8-import-order
flake8-blind-except==0.1.1  # https://github.com/elijahandrews/flake8-blind-except
flake8-bugbear==19.8.0      # https://github.com/PyCQA/flake8-bugbear
flake8-django==0.0.4        # https://github.com/rocioar/flake8-django
flake8-debugger==3.1.0      # https://github.com/JBKahn/flake8-debugger
flake8-broken-line==0.1.1   # https://github.com/sobolevn/flake8-broken-line
flake8-print==3.1.0         # https://pypi.org/project/flake8print/
flake8-eradicate==0.2.1     # https://github.com/sobolevn/flake8-eradicate
flake8-quotes==2.1.0        # https://github.com/zheller/flake8-quotes
flake8-string-format==0.2.3 # https://github.com/xZise/flake8-string-format
flake8-mutable==1.2.0       # https://github.com/ebeweber/flake8-mutable
flake8-docstrings==1.4.0    # https://github.com/PyCQA/flake8-docstrings
flake8-isort==2.8.0

